modularized addBook and removeBook
