+++ index.js
@@ -1,3 +1,4 @@
+/* eslint-disable */
 import BookList from './modules/book-list.js';
 
 import {
@@ -19,5 +20,4 @@ const bookListInstance = new BookList();
 Booklink.addEventListener('click', menu.showList);
 Addlink.addEventListener('click', menu.showAddBook);
 Contactlink.addEventListener('click', menu.showContact);
-const bookToRemove = bookListInstance.books[0];
-bookListInstance.removeBook(bookToRemove);
+
