+++ modules/book-list.js
@@ -1,5 +1,4 @@
-import Book from './book.js';
-
+import { addBook, removeBook } from './bookUtils.js'
 class BookList {
   constructor() {
     this.books = [];
@@ -8,28 +7,7 @@ class BookList {
   }
 
   addBook = () => {
-    const addBtn = document.getElementById('add-btn');
-    addBtn.addEventListener('click', () => {
-      const storedTitle = localStorage.getItem('title');
-      const storedAuthor = localStorage.getItem('author');
-
-      if (storedTitle && storedAuthor) {
-        document.getElementById('title').value = storedTitle;
-        document.getElementById('author').value = storedAuthor;
-      }
-
-      const titleInput = document.getElementById('title').value;
-      const authorInput = document.getElementById('author').value;
-
-      const book = new Book(titleInput, authorInput);
-      this.books.push(book);
-
-      localStorage.setItem('title', '');
-      localStorage.setItem('author', '');
-      localStorage.setItem('books', JSON.stringify(this.books));
-
-      this.displayBook(book);
-    });
+    addBook(this); 
   }
 
   loadBooks = () => {
@@ -66,12 +44,9 @@ class BookList {
   }
 
   removeBook = (book) => {
-    const index = this.books.indexOf(book);
-    if (index !== -1) {
-      this.books.splice(index, 1);
-      localStorage.setItem('books', JSON.stringify(this.books));
-    }
+    removeBook(this, book);
   }
+
 }
 
 export default BookList;
